agents/rag-agent/requirements.txt
langchain>=1.2.10
langchain-core>=1.2.17
langchain-ollama>=1.0.1
ddgs>=8.0.0
beautifulsoup4>=4.14.0
lxml>=6.0.0
requests>=2.32.0
python-dotenv>=1.0.0
